+++ api/src/identity-access-management/application/user/user.route.js
@@ -1,6 +1,12 @@
 import Joi from 'joi';
+import XRegExp from 'xregexp';
 
+import { userVerification } from '../../../../lib/application/preHandlers/user-existence-verification.js';
+import { config } from '../../../shared/config.js';
+import { identifiersType } from '../../../shared/domain/types/identifiers-type.js';
 import { userController } from './user.controller.js';
+
+const { passwordValidationPattern } = config.account;
 
 export const userRoutes = [
   {
@@ -22,7 +28,41 @@ export const userRoutes = [
         },
       },
       handler: (request, h) => userController.save(request, h),
-      tags: ['api', 'user-account'],
+      tags: ['identity-access-management', 'api', 'user'],
+    },
+  },
+  {
+    method: 'PATCH',
+    path: '/api/users/{id}/password-update',
+    config: {
+      auth: false,
+      pre: [
+        {
+          method: (request, h) => userVerification.verifyById(request, h),
+          assign: 'user',
+        },
+      ],
+      handler: (request, h) => userController.updatePassword(request, h),
+      validate: {
+        options: {
+          allowUnknown: true,
+        },
+        params: Joi.object({
+          id: identifiersType.userId,
+        }),
+        payload: Joi.object({
+          data: {
+            attributes: {
+              password: Joi.string().pattern(XRegExp(passwordValidationPattern)).required(),
+            },
+          },
+        }),
+      },
+      notes: [
+        "- Met à jour le mot de passe d'un utilisateur identifié par son id\n" +
+          "- Une clé d'identification temporaire permet de vérifier l'identité du demandeur",
+      ],
+      tags: ['identity-access-managements', 'api', 'user'],
     },
   },
 ];
